change prompting for unknown errors on download.
