+++ src/js/helpers/FileConversionHelpers/index.js
@@ -2,6 +2,7 @@
 import React from 'react';
 import path from 'path-extra';
 // helpers
+import * as REPO from '../../helpers/Repo.js';
 import * as UsfmFileConversionHelpers from './UsfmFileConversionHelpers';
 import * as ZipFileConversionHelpers from './ZipFileConversionHelpers';
 
@@ -70,3 +71,58 @@ export const getSafeErrorMessage = (error, defaultErrorMessage) => {
   return errorMessage;
 };
 
+/**
+ * convert git error message to localized message and determine if known or unknown
+ * @param {String|Object} error
+ * @param {Function} translate
+ * @param {String} projectName
+ * @return {{isUnknown: boolean, message: string}}
+ */
+export function getLocalizedErrorMessage(error, translate, projectName) {
+  let message = 'unknown';
+  let isUnknown = true;
+
+  if (error.status === 401) {
+    message = translate('users.session_invalid');
+    isUnknown = false;
+  } else {
+    const errorStr = error.toString();
+
+    if (errorStr.includes(REPO.GIT_ERROR_REPO_ARCHIVED)) {
+      message = translate('projects.archived',
+        {
+          project_name: projectName,
+          door43: translate('_.door43'),
+          app_name: translate('_.app_name'),
+        });
+      isUnknown = false;
+    } else if (error.code === REPO.NETWORK_ERROR_IP_ADDR_NOT_FOUND ||
+      errorStr.includes(REPO.GIT_ERROR_UNABLE_TO_CONNECT) ||
+      errorStr.includes(REPO.NETWORK_ERROR_TIMEOUT) ||
+      errorStr.includes(REPO.NETWORK_ERROR_INTERNET_DISCONNECTED) ||
+      errorStr.includes(REPO.NETWORK_ERROR_UNABLE_TO_ACCESS) ||
+      errorStr.includes(REPO.NETWORK_ERROR_REMOTE_HUNG_UP)) {
+      message = translate('no_internet');
+      isUnknown = false;
+    } else if (errorStr.includes(REPO.GIT_ERROR_PUSH_NOT_FF)) {
+      message = translate('projects.upload_modified_error',
+        { project_name: projectName, door43: translate('_.door43') });
+      isUnknown = false;
+    } else if (errorStr.includes(REPO.GIT_ERROR_UNKNOWN_PROBLEM)) {
+      const parts = errorStr.split(REPO.GIT_ERROR_UNKNOWN_PROBLEM);
+      let details = parts.length > 1 ? parts[1] : errorStr;
+
+      if (details[0] === ':') {
+        details = details.substr(1).trim();
+      }
+      console.error(`Unknown GIT error: ${details}`);
+    } else if (error.hasOwnProperty('message')) {
+      console.error(`Unknown error: ${error.message}`);
+    } else if (error.hasOwnProperty('data') && error.data) {
+      console.error(`Unknown upload error: ${error.data}`);
+    } else { // unknown error
+      console.error('Unknown error:', errorStr || error);
+    }
+  }
+  return { message, isUnknown };
+}
